pyproject.toml
[project]
name = "fuel-route-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "geopy>=2.4.1",
    "haversine>=2.9.0",
    "polyline>=2.0.4",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
